backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
spacy
sqlalchemy
aiosqlite
pydantic-settings
httpx
beautifulsoup4
python-magic
pytest
pytest-asyncio
ruff
mypy
python-json-logger
slowapi
